move get_values to mainclasses
